agents/a2a_agents/git_agent/pyproject.toml
[project]
name = "agent-2-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk==0.3.0",
    "aiohttp>=3.12.15",
    "autogen-agentchat>=0.7.4",
    "autogen-ext[openai]>=0.7.4",
    "git-python>=1.0.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]
